refactor: 인터셉터 관련 설정 분리
