+++ src/main/java/com/spaceclub/club/controller/ClubController.java
@@ -6,7 +6,7 @@ import com.spaceclub.club.controller.dto.ClubUpdateRequest;
 import com.spaceclub.club.domain.Club;
 import com.spaceclub.club.service.ClubService;
 import com.spaceclub.global.Authenticated;
-import com.spaceclub.global.config.s3.properties.S3Properties;
+import com.spaceclub.global.config.s3.S3Properties;
 import com.spaceclub.global.jwt.vo.JwtUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
