+++ src/main/java/com/spaceclub/global/config/s3/S3Properties.java
@@ -1,4 +1,4 @@
-package com.spaceclub.global.config.s3.properties;
+package com.spaceclub.global.config.s3;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
