+++ src/main/java/com/spaceclub/global/interceptor/AuthorizationInterceptor.java
@@ -1,5 +1,7 @@
 package com.spaceclub.global.interceptor;
 
+import com.spaceclub.global.config.InterceptorProperties;
+import com.spaceclub.global.config.InterceptorProperties.pathMethod;
 import com.spaceclub.global.exception.TokenException;
 import com.spaceclub.global.jwt.Claims;
 import com.spaceclub.global.jwt.JwtManager;
@@ -12,33 +14,29 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import static com.spaceclub.global.exception.GlobalExceptionCode.BAD_REQUEST;
 import static java.util.Collections.unmodifiableSet;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static org.springframework.http.HttpMethod.GET;
-import static org.springframework.http.HttpMethod.POST;
 
 @Component
 @RequiredArgsConstructor
 public class AuthorizationInterceptor implements HandlerInterceptor {
 
-    private static final String TOKEN_PREFIX = "Bearer ";
-
-    private static final Map<String, HttpMethod> PATH_TO_EXCLUDE = Map.of(
-            "/api/v1/users", POST,
-            "/api/v1/clubs/invites/", GET,
-            "/api/v1/events", GET,
-            "/api/v1/events/searches", GET,
-            "/api/v1/users/oauth", POST
-    );
+    private final InterceptorProperties interceptorProperties;
 
     private final AccountService accountService;
 
     private final JwtManager jwtManager;
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        System.out.println("interceptorProperties.tokenPrefix() = " + interceptorProperties.tokenPrefix());
+
+        Map<String, HttpMethod> PATH_TO_EXCLUDE = interceptorProperties.pathToExclude().stream()
+                .collect(Collectors.toMap(pathMethod::path, pathMethod::method));
+
         for (Map.Entry<String, HttpMethod> entry : unmodifiableSet(PATH_TO_EXCLUDE.entrySet())) {
             boolean matchesPathAndMethod = request.getRequestURI().contains(entry.getKey()) &&
                     request.getMethod().equals(entry.getValue().name());
@@ -51,7 +49,7 @@ public class AuthorizationInterceptor implements HandlerInterceptor {
         String authorizationHeader = request.getHeader(AUTHORIZATION);
         validateHeader(authorizationHeader);
 
-        String header = authorizationHeader.replace(TOKEN_PREFIX, "");
+        String header = authorizationHeader.replace(interceptorProperties.tokenPrefix(), "");
         Claims claims = jwtManager.getClaims(header);
 
         return accountService.isAuthenticatedUser(claims.getId(), claims.getUsername());
