+++ src/main/java/com/spaceclub/user/service/UserService.java
@@ -2,7 +2,7 @@ package com.spaceclub.user.service;
 
 import com.spaceclub.global.config.s3.S3Folder;
 import com.spaceclub.global.config.s3.S3ImageUploader;
-import com.spaceclub.global.config.s3.properties.S3Properties;
+import com.spaceclub.global.config.s3.S3Properties;
 import com.spaceclub.user.domain.User;
 import com.spaceclub.user.repository.UserRepository;
 import com.spaceclub.user.service.vo.RequiredProfile;
