+++ src/main/resources/application.yml
@@ -20,11 +20,19 @@ spring:
   config:
     activate:
       on-profile: develop
-    import: develop/application-develop.yml, global/application-oauth.yml, global/application-s3.yml
+    import:
+      - develop/application-develop.yml
+      - global/application-oauth.yml
+      - global/application-s3.yml
+      - global/application-config.yml
 
 ---
 spring:
   config:
     activate:
       on-profile: local
-    import: local/application-local.yml, global/application-oauth.yml, global/application-s3.yml
+    import:
+      - local/application-local.yml
+      - global/application-oauth.yml
+      - global/application-s3.yml
+      - global/application-config.yml
